# Improve Enum.TryParse performance. Fixes #70