@AlexGhiondea 2015-02-04
Thanks for the contribution.

Since this code is shared with desktop we need to make sure we don't break compatibility with it. 

Cases where the behavior is different before/after the change are  considered breaking changes and need to be mitigated. 

There are a couple of ways we mitigate breaking changes:
- the code can be if-def'ed to only impact CoreCLR
- we can code the fix such that it behaves one way on older platforms (ie. current version of Desktop) but get the new behavior for newer platforms. (this is handled by/with the AppContext class).

For this specific case, you can probably special case the situation where the enum backing type is smaller than the value that you need to assign to it and in that case have the old behavior.