@Alexx999 2015-02-05
OK, I got it - we play safe here :)
I've got one idea how to make it behave exactly like the old one without hurting performance too much, I'll make it work that way