@Alexx999 2015-02-05
Implemented that change.
New performance figures:
```
Good run: 214ms
Bad run: 187ms
Big numbers run: 443ms
Small numbers run: 319ms
All letters run: 121ms
```
It is now a bit slower with large numbers (still 25 times faster than original) and has roughly equal performance with small numbers.