@AlexGhiondea 2015-02-06
Thanks for the contribuition. I am going to run some validation and if everything is ok I'll merge this in.