@nguerrera 2015-02-06
Note that bool is a valid enum underlying type. How does this change interact with that? 

e.g. Given:
```
.class public sealed PeculiarEnum extends [mscorlib]System.Enum
{
  .field public specialname rtspecialname bool value__
  .field public static literal valuetype PeculiarEnum A = bool(false)
  .field public static literal valuetype PeculiarEnum B = bool(true)
}
```
Does the behavior of  `Enum.Parse("0", typeof(PeculiarEnum))` change?

At a glance, I suspect it would have thrown FormatException before, but return PeculiarEnum.A now.