@Alexx999 2015-02-06
@nguerrera yeah, you're right.
I've updated test with this particular case - I think we should update Enum tests in CoreFx too.

I think this particular case can be worked around with additional check.
```
Good run: 213ms
Bad run: 182ms
Big numbers run: 482ms
Small numbers run: 356ms
All letters run: 122ms
```