@Alexx999 2015-02-06
Updated test to check if char-based enum works correctly.
It turns out that in code
```C#
Enum.Parse("0", typeof(CharEnum))
```
"0" was treated like a char and not like number, so the result must be
```C#
Enum.ToObject(typeof(CharEnum), '0')
```
and it was equal to
```C#
Enum.ToObject(typeof(CharEnum), 0)
```

Now new code basically mimics behavior of Convert.ChangeType approach, but without throwing.
As a bonus - it's now faster compared to previous solution with extra checks.

```
Good run: 214ms
Bad run: 203ms
Big numbers run: 247ms
Small numbers run: 229ms
All letters run: 124ms
```