@AlexGhiondea 2015-02-06
Overall the change looks good. I left some comments in the PR. 
Could you squash the commits into a single one?

If nothing comes out of the test run I am doing I'll merge this in!