@Alexx999 2015-02-06
@AlexGhiondea OK, I cleaned up remaining things and squashed it to one commit.