@AlexGhiondea 2015-02-07
Thanks for all the work you did on this change. I have kicked off a different test run and should have results on Monday. The test run I did didn't show up any issues.

However, I talked with @jkotas about this and, to be on the safe side, we should update the code so that in the default case of the switch statement we fall back to the slow behavior we had before. This would cover the bool/float/double case.

Can you also check what happens when we have an enum backed by a float/double? I know you cannot create these in C#, but via IL or C++ they could still be generated and we need to make sure we can correctly handle them.

Again, thank you for doing the due diligence around this fix.