@Alexx999 2015-02-07
@AlexGhiondea Indeed, @jkotas is right - there is possibility of having enum backed by float/double/IntPtr/UIntPtr.
This can be seen here:
https://github.com/dotnet/coreclr/blob/9df15b9fdfb55ce2076a5e7eede8ca71481e5e3b/src/mscorlib/src/System/Enum.cs#L749
But on the same time ToObject() method doesn't support any of those types (as opposed to bool and char - they're actually supported) so parsing it will fail anyway, see
https://github.com/dotnet/coreclr/blob/9df15b9fdfb55ce2076a5e7eede8ca71481e5e3b/src/mscorlib/src/System/Enum.cs#L550