@Alexx999 2015-02-07
Okay, added one more fixup that will ensure correct behavior in all cases - it will throw ArgumentException with same message as before.
Performance seems to be unaffected, at least difference is less than measurement error.
```
Good run: 212ms
Bad run: 204ms
Big numbers run: 245ms
Small numbers run: 237ms
All letters run: 129ms
```