@AlexGhiondea 2015-02-10
@Alexx999 I would like to thank you for the work you put into this PR.

We have updated our contribution guidelines to better call out the bar for accepting pull requests. You can find it here: https://github.com/dotnet/coreclr/wiki/Contribution-guidelines

We are aware of the work you put into this PR and we truly appreciate it! But, unfortunately, since it is not a mainline scenario and has unknown risk, this PR does not meet the bar.

We are sorry for not being able to accept it and hope you will continue making contributions!

Thank you again for your contribution!