@Phrohdoh 2015-02-11
It is quite sad to see a PR like this being discarded after all of the work put forward here.