@Alexx999 2015-02-11
that's a bit of surprise - it just seemed like all possible cases are covered at last.

but probably you are right - this case is not really common.