@AlexGhiondea 2015-02-11
@Phrohdoh I know.

One thing I would like to do is, with @Alexx999's permission, capture the tests in the CoreFX repo (I can make the change or review a PR). 
That way, if we ever touch that area in the future we have a better understanding of the tricky parts.