@weshaggard 2015-02-11
@Phrohdoh I agree it is unfortunate as well and I also appreciate the effort that @Alexx999 put into this and while we were having the conversation on the PR we were also having a conversation internally to try and figure out the appropriate bar, hopefully that is a little clearer now. In this case while it would be nice to take the reward isn't worth the potential risk. 

Thanks again @alex999 and I look forward to your future contributions.